// projects/app/src/app/interfaces/oanda.ts
export interface AccountProperty {
  /** The Accounts identifier */
  id: string;

  /**
   * The Account's associated MT4 Account ID. This field will not be present
   * if the Account is not an MT4 account.
   */
  mt4AccountID?: number;

  /** The Accounts tags */
  tags: string[];
}
/** A tag associated with an entity. */
interface Tag {
  /**The type of the tag. */
  type: string;

  /**The name of the tag. */
  name: string;
}

export interface OandaError {
  lastTransactionID: string;
  errorMessage: string;
  errorCode: string;
}

/** The type of an Instrument. */
export type InstrumentType = 'CURRENCY' | 'CFD' | 'METAL';
// ...
type GuaranteedStopLossOrderModeForInstrument =
  | 'DISABLED'
  | 'ALLOWED'
  | 'REQUIRED';
// ...
export interface Instrument {
  /**
   * The name of the Instrument
   */
  name: string;

  /**
   * The type of the Instrument
   */
  type: InstrumentType;

  /**
   * The display name of the Instrument
   */
  displayName: string;

  /**
   * The location of the “pip” for this instrument. The decimal position of
   * the pip in this Instrument's price can be found at 10 ^ pipLocation e.g.
   * -4 pipLocation results in a decimal pip position of 10 ^ -4 = 0.0001.
   */
  pipLocation: number;

  /**
   * The number of decimal places that should be used to display prices for
   * this instrument. e.g. a displayPrecision of 5 would result in a price of
   * “1” being displayed as “1.00000”
   */
  displayPrecision: number;

  /**
   * The amount of decimal places that may be provided when specifying the
   * number of units traded for this instrument.
   */
  tradeUnitsPrecision: number;

  /**
   * The smallest number of units allowed to be traded for this instrument.
   */
  minimumTradeSize: string;

  /**
   * The maximum trailing stop distance allowed for a trailing stop loss
   * created for this instrument. Specified in price units.
   */
  maximumTrailingStopDistance: string;

  /**
   * The minimum distance allowed between the Trade's fill price and the
   * configured price for guaranteed Stop Loss Orders created for this
   * instrument. Specified in price units.
   */
  minimumGuaranteedStopLossDistance: string;

  /**
   * The minimum trailing stop distance allowed for a trailing stop loss
   * created for this instrument. Specified in price units.
   */
  minimumTrailingStopDistance: string;

  /**
   * The maximum position size allowed for this instrument. Specified in
   * units.
   */
  maximumPositionSize: string;

  /**
   * The maximum units allowed for an Order placed for this instrument.
   * Specified in units.
   */
  maximumOrderUnits: string;

  /**
   * The margin rate for this instrument.
   */
  marginRate: string;

  /**
   * The commission structure for this instrument.
   */
  commission: InstrumentCommission;

  /**
   * The current Guaranteed Stop Loss Order mode of the Account for this
   * Instrument.
   */
  guaranteedStopLossOrderMode: GuaranteedStopLossOrderModeForInstrument;

  /**
   * The amount that is charged to the account if a guaranteed Stop Loss Order
   * is triggered and filled. The value is in price units and is charged for
   * each unit of the Trade. This field will only be present if the Account's
   * guaranteedStopLossOrderMode for this Instrument is not ‘DISABLED'.
   */
  guaranteedStopLossOrderExecutionPremium: string;

  /**
   * The guaranteed Stop Loss Order level restriction for this instrument.
   * This field will only be present if the Account's
   * guaranteedStopLossOrderMode for this Instrument is not ‘DISABLED'.
   */
  guaranteedStopLossOrderLevelRestriction: GuaranteedStopLossOrderLevelRestriction;

  /**
   * Financing data for this instrument.
   */
  financing: InstrumentFinancing;

  /**
   * The tags associated with this instrument.
   */
  tags: Tag[];
}
// ...
export type PositionStateFilter = 'ALL' | 'OPEN';
// ...
export interface TradeSummary {
  /** The Trade's identifier, unique within the Trade's Account. */
  id: string;

  /** The Trade's Instrument. */
  instrument: string;

  /** The execution price of the Trade. */
  price: string;

  /** The date/time when the Trade was opened. */
  openTime: string;

  /** The current state of the Trade. */
  state: TradeState;

  /**
   * The initial size of the Trade. Negative values indicate a short Trade,
   * and positive values indicate a long Trade.
   */
  initialUnits: string;

  /**
   * The margin required at the time the Trade was created. Note, this is the
   * ‘pure' margin required, it is not the ‘effective' margin used that
   * factors in the trade risk if a GSLO is attached to the trade.
   */
  initialMarginRequired: string;

  /**
   * The number of units currently open for the Trade. This value is reduced
   * to 0.0 as the Trade is closed.
   */
  currentUnits: string;

  /** The total profit/loss realized on the closed portion of the Trade. */
  realizedPL: string;

  /** The unrealized profit/loss on the open portion of the Trade. */
  unrealizedPL: string;

  /** Margin currently used by the Trade. */
  marginUsed: string;

  /**
   * The average closing price of the Trade. Only present if the Trade has
   * been closed or reduced at least once.
   */
  averageClosePrice: string;

  /**
   * The IDs of the Transactions that have closed portions of this Trade.
   */
  closingTransactionIDs: string[];

  /** The financing paid/collected for this Trade. */
  financing: string;

  /** The dividend adjustment paid for this Trade. */
  dividendAdjustment: string;

  /**
   * The date/time when the Trade was fully closed. Only provided for Trades
   * whose state is CLOSED.
   */
  closeTime: string;

  /** The client extensions of the Trade. */
  clientExtensions: ClientExtensions;

  /**
   * ID of the Trade's Take Profit Order, only provided if such an Order
   * exists.
   */
  takeProfitOrderID: string;

  /**
   * ID of the Trade's Stop Loss Order, only provided if such an Order exists.
   */
  stopLossOrderID: string;

  /**
   * ID of the Trade's Guaranteed Stop Loss Order, only provided if such an
   * Order exists.
   */
  guaranteedStopLossOrderID: string;

  /**
   * ID of the Trade's Trailing Stop Loss Order, only provided if such an
   * Order exists.
   */
  trailingStopLossOrderID: string;
}
// ...
interface GuaranteedStopLossOrderParameters {
  /**
   * The current guaranteed Stop Loss Order mutability setting of the Account
   * when market is open.
   */
  mutabilityMarketOpen: GuaranteedStopLossOrderMutability;

  /**
   * The current guaranteed Stop Loss Order mutability setting of the Account
   * when market is halted.
   */
  mutabilityMarketHalted: GuaranteedStopLossOrderMutability;
}

/**
 * For Accounts that support guaranteed Stop Loss Orders, describes the actions
 * that can be be performed on guaranteed Stop Loss Orders.
 */
type GuaranteedStopLossOrderMutability =
  | 'FIXED'
  | 'REPLACEABLE'
  | 'CANCELABLE'
  | 'PRICE_WIDEN_ONLY';

/**
 * The overall behaviour of the Account regarding guaranteed Stop Loss Orders.
 */
type GuaranteedStopLossOrderMode = 'DISABLED' | 'ALLOWED' | 'REQUIRED';

export interface AccountSummaryResponse {
  /**
   * The summary of the requested Account.
   */
  account: AccountSummary;

  /**
   * The ID of the most recent Transaction created for the Account.
   */
  lastTransactionID: string;
}

/**
 * The full details of a client's Account. This includes full open Trade, open
 * Position and pending Order representation.
 */
export interface Account {
  /**
   * The Account's identifier
   */
  id: string;

  /**
   * Client-assigned alias for the Account. Only provided if the Account has
   * an alias set
   */
  alias: string;

  /**
   * The home currency of the Account
   */
  currency: string;

  /**
   * ID of the user that created the Account.
   */
  createdByUserID: number;

  /**
   * The date/time when the Account was created.
   */
  createdTime: string;

  /**
   * The current guaranteed Stop Loss Order settings of the Account. This
   * field will only be present if the guaranteedStopLossOrderMode is not
   * ‘DISABLED'.
   */
  guaranteedStopLossOrderParameters: GuaranteedStopLossOrderParameters;

  /**
   * The current guaranteed Stop Loss Order mode of the Account.
   */
  guaranteedStopLossOrderMode: GuaranteedStopLossOrderMode;

  /**
   * The date/time that the Account's resettablePL was last reset.
   */
  resettablePLTime: string;

  /**
   * Client-provided margin rate override for the Account. The effective
   * margin rate of the Account is the lesser of this value and the OANDA
   * margin rate for the Account's division. This value is only provided if a
   * margin rate override exists for the Account.
   */
  marginRate: string;

  /**
   * The number of Trades currently open in the Account.
   */
  openTradeCount: number;

  /**
   * The number of Positions currently open in the Account.
   */
  openPositionCount: number;

  /**
   * The number of Orders currently pending in the Account.
   */
  pendingOrderCount: number;

  /**
   * Flag indicating that the Account has hedging enabled.
   */
  hedgingEnabled: boolean;

  /**
   * The total unrealized profit/loss for all Trades currently open in the
   * Account.
   */
  unrealizedPL: string;

  /**
   * The net asset value of the Account. Equal to Account balance +
   * unrealizedPL.
   */
  NAV: string;

  /**
   * Margin currently used for the Account.
   */
  marginUsed: string;

  /**
   * Margin available for Account currency.
   */
  marginAvailable: string;

  /**
   * The value of the Account's open positions represented in the Account's
   * home currency.
   */
  positionValue: string;

  /**
   * The Account's margin closeout unrealized PL.
   */
  marginCloseoutUnrealizedPL: string;

  /**
   * The Account's margin closeout NAV.
   */
  marginCloseoutNAV: string;

  /**
   * The Account's margin closeout margin used.
   */
  marginCloseoutMarginUsed: string;

  /**
   * The Account's margin closeout percentage. When this value is 1.0 or above
   * the Account is in a margin closeout situation.
   */
  marginCloseoutPercent: string;

  /**
   * The value of the Account's open positions as used for margin closeout
   * calculations represented in the Account's home currency.
   */
  marginCloseoutPositionValue: string;

  /**
   * The current WithdrawalLimit for the account which will be zero or a
   * positive value indicating how much can be withdrawn from the account.
   */
  withdrawalLimit: string;

  /**
   * The Account's margin call margin used.
   */
  marginCallMarginUsed: string;

  /**
   * The Account's margin call percentage. When this value is 1.0 or above the
   * Account is in a margin call situation.
   */
  marginCallPercent: string;

  /**
   * The current balance of the account.
   */
  balance: string;

  /**
   * The total profit/loss realized over the lifetime of the Account.
   */
  pl: string;

  /**
   * The total realized profit/loss for the account since it was last reset by
   * the client.
   */
  resettablePL: string;

  /**
   * The total amount of financing paid/collected over the lifetime of the
   * account.
   */
  financing: string;

  /**
   * The total amount of commission paid over the lifetime of the Account.
   */
  commission: string;

  /**
   * The total amount of dividend adjustment paid over the lifetime of the
   * Account in the Account's home currency.
   */
  dividendAdjustment: string;

  /**
   * The total amount of fees charged over the lifetime of the Account for the
   * execution of guaranteed Stop Loss Orders.
   */
  guaranteedExecutionFees: string;

  /**
   * The date/time when the Account entered a margin call state. Only provided
   * if the Account is in a margin call.
   */
  marginCallEnterTime: string;

  /**
   * The number of times that the Account's current margin call was extended.
   */
  marginCallExtensionCount: number;

  /**
   * The date/time of the Account's last margin call extension.
   */
  lastMarginCallExtensionTime: string;

  /**
   * The ID of the last Transaction created for the Account.
   */
  lastTransactionID: string;

  /**
   * The details of the Trades currently open in the Account.
   */
  trades: TradeSummary[];

  /**
   * The details all Account Positions.
   */
  positions: Position[];

  /**
   * The details of the Orders currently pending in the Account.
   */
  orders: Order[];
}
// ...
export interface AccountPoll {
  /**
   * The changes to the Account’s Orders, Trades and Positions since the
   * specified Transaction ID. Only provided if the sinceTransactionID is
   * supplied to the poll request.
   */
  changes: AccountChanges;

  /** The Account’s current price-dependent state. */
  state: AccountChangesState;

  /**
   * The ID of the last Transaction created for the Account.  This Transaction
   * ID should be used for future poll requests, as the client has already
   * observed all changes up to and including it.
   */
  lastTransactionID: string;
}